Tests: separate rollup and shrink action tests in own classes
This separates the ILM rollup and shrink actions tests from TimeSeriesLifecycleActionsIT
into their own classes.
